+++ lib/pages/home/widgets/drawer.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:flutter_svg/flutter_svg.dart';
+import 'package:flutter_svg_plus/svg.dart';
+import 'package:in_app_review/in_app_review.dart';
 import 'package:memorize_scripture/common/strings.dart';
 import 'package:memorize_scripture/pages/about/about_page.dart';
 import 'package:memorize_scripture/pages/account/account_page.dart';
@@ -63,6 +64,12 @@ class MenuDrawer extends StatelessWidget {
                     builder: (_) => const AccountPage(),
                   ),
                 );
+              },
+            ),
+            ListTile(
+              title: const Text('Rate this app'),
+              onTap: () async {
+                await InAppReview.instance.requestReview();
               },
             ),
             ListTile(